requirements.txt
langchain==0.1.9
langchain-openai==0.0.8
openai==0.28.0
playwright==1.41.2
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
gradio==4.19.2
asyncio==3.4.3
